Build everything as one type of BUILD.
We need to build everything with the same compiler as source-only libs
can depend on not-source-only libs. And hence the not-source-only libs
might not exist if we use different BUILD vars for the two types.

This also moves from using an explicit BUILD setting to using the last
part of the branch name "TYPE/VERSION/BUILD".